--- source/api_c/include/cxx/README.md
@@ -1,5 +1,0 @@
-# Header-only C++ 11 library based on C library
-
-This head-only C++ 11 library is ONLY dependent on the C library.
-
-These headers should not be mixed with headers from the C++ library.
